tests: fix 17-cilium_policy_id-remove
17-cilium_policy_id-remove used invalid container names and was silently
failing. Added `set -e` in the test to ensure that the test fails
when any command returns a non-zero exit code. Also added log messages
to commentate what the test is doing so it is more clear to a user
unfamiliar with the test.

Signed-off by: Ian Vernon <ian@cilium.io>